adding unittests for thresholding function
